move postgres fields tests to separate folder similar to DRF
