+++ test/input/func_noerror_model_fields.py
@@ -89,53 +89,3 @@ class LotsOfFieldsModel(models.Model):
         # print(self.uuidfield.int)  # Don't know how to properly check this one
         print(self.uuidfield.variant)
         print(self.uuidfield.version)
-
-
-try:
-    # Imports psycopg2, which might not be installed
-    from django.contrib.postgres import fields
-except ImportError:
-    pass
-else:
-    class PostgresFieldsModel(models.Model):
-        arrayfield = fields.ArrayField(models.CharField())
-        hstorefield = fields.HStoreField()
-        jsonfield = fields.JSONField()
-        rangefield = fields.RangeField()
-        integerrangefield = fields.IntegerRangeField()
-        bigintegerrangefield = fields.BigIntegerRangeField()
-        floatrangefield = fields.FloatRangeField()
-        datetimerangefield = fields.DateTimeRangeField()
-        daterangefield = fields.DateRangeField()
-
-        def arrayfield_tests(self):
-            sorted_array = self.arrayfield.sort()
-            print(sorted_array)
-
-        def dictfield_tests(self):
-            print(self.hstorefield.keys())
-            print(self.hstorefield.values())
-            print(self.hstorefield.update({'foo': 'bar'}))
-
-            print(self.jsonfield.keys())
-            print(self.jsonfield.values())
-            print(self.jsonfield.update({'foo': 'bar'}))
-
-        def rangefield_tests(self):
-            print(self.rangefield.lower)
-            print(self.rangefield.upper)
-
-            print(self.integerrangefield.lower)
-            print(self.integerrangefield.upper)
-
-            print(self.bigintegerrangefield.lower)
-            print(self.bigintegerrangefield.upper)
-
-            print(self.floatrangefield.lower)
-            print(self.floatrangefield.upper)
-
-            print(self.datetimerangefield.lower)
-            print(self.datetimerangefield.upper)
-
-            print(self.daterangefield.lower)
-            print(self.daterangefield.upper)
